requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-docx==0.8.11
docx2txt==0.8
langchain
faiss-cpu==1.7.4
langchain-community
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain-experimental
PyJWT==2.8.0
fastapi==0.104.1
pydantic
pyvis==0.3.2
pypdf
pymongo
langchain_elasticsearch
elasticsearch
torch
bert-extractive-summarizer
